Updated code

+++ src/com/cret/gui/ProjectListController.java
@@ -1,17 +1,15 @@
 package com.cret.gui;
 
 import java.io.IOException;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Optional;
 import com.cret.db.Database;
-import com.cret.db.DbUtils;
 import com.cret.projects.Project;
-import com.cret.staticData.structures;
+import com.cret.staticdata.structures;
+
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -32,7 +30,7 @@ import javafx.scene.control.MenuItem;
  *
  */
 public class ProjectListController {
-	
+
 	/**
 	 * Database object.
 	 */
@@ -45,7 +43,7 @@ public class ProjectListController {
 	 */
 	@FXML
 	private Button btnOpenProject;
-	
+
 	/**
 	 * FXML Element
 	 * 
@@ -53,7 +51,7 @@ public class ProjectListController {
 	 */
 	@FXML
 	private Button btnCancel;
-	
+
 	/**
 	 * FXML Element
 	 * 
@@ -61,7 +59,7 @@ public class ProjectListController {
 	 */
 	@FXML
 	private Button btnDeleteProject;
-	
+
 	/**
 	 * FXML Element
 	 * 
@@ -69,7 +67,7 @@ public class ProjectListController {
 	 */
 	@FXML
 	private ListView<String> listProjects;
-	
+
 	/**
 	 * FXML Element
 	 * 
@@ -77,7 +75,7 @@ public class ProjectListController {
 	 */
 	@FXML
 	private MenuItem btnImport;
-	
+
 	/**
 	 * FXML Element
 	 * 
@@ -85,7 +83,6 @@ public class ProjectListController {
 	 */
 	@FXML
 	private MenuItem btnExport;
-	
 
 	/**
 	 * FXML Function
@@ -96,77 +93,71 @@ public class ProjectListController {
 	 */
 	@FXML
 	private void closeWindow(ActionEvent event) {
-		
-	    Stage stage = (Stage) btnCancel.getScene().getWindow();
-	    stage.close();
+
+		db.disconnect();
+		Stage stage = (Stage) btnCancel.getScene().getWindow();
+		stage.close();
 	}
-	
+
 	/**
 	 * FXML Function
 	 * 
 	 * Opens selected project.
 	 * 
 	 * @param event Click event of the user.
+	 * @throws SQLException
 	 */
 	@FXML
-	private void openProject(ActionEvent event) {
-		
+	private void openProject(ActionEvent event) throws SQLException {
+
 		/**
 		 * Reset all structures.
 		 */
-		structures.resetDashboardStructures();
-		structures.resetAnalisysStructures();
+		structures.resetDashboardStructures1();
+		structures.resetAnalisysStructures1();
 		structures.resetRAWStructures();
-		
-		//Get selected project.
+
+		structures.valuesInDashboard1 = new HashMap<>();
+		// Get selected project.
 		String projectName = listProjects.getSelectionModel().getSelectedItem().toString();
-		
+
 		/**
 		 * Set mode to dashboard
 		 */
-		structures.mode = "Dashboard";
-		
-		structures.valuesInDashboard = new HashMap<>();
-		
-        
-        String sqlQuery = "SELECT name, ID, byte FROM data WHERE projectName = ?";
-        
-        try (Connection conn = DbUtils.connect();
+		structures.modeInterface1 = "Dashboard";
 
-        	 PreparedStatement pstmt  = conn.prepareStatement(sqlQuery)){
-        	            pstmt.setString(1,projectName);
-        	            ResultSet rs  = pstmt.executeQuery();
-            while (rs.next()) {
-            	
-            	String dataName = rs.getString("name");
-            	String id = rs.getString("id");
-            	String byteNumber = rs.getString("byte");
-                
-                if (!structures.valuesInDashboard.containsKey(dataName)) {
-                	structures.valuesInDashboard.put(dataName, new ArrayList<>());
-                }
-                
-                structures.valuesInDashboard.get(dataName).add(id);
-                structures.valuesInDashboard.get(dataName).add(byteNumber);
-                structures.dataQ1Dash = new HashMap<>();
-                
-            }
-            
-            //Sets new project
-            RootViewController.projectObj = new Project(projectName);
-            RootViewController.projectName.setValue(projectName);
-            
-            //Closes window
-    	    Stage stage = (Stage) btnCancel.getScene().getWindow();
-    	    stage.close();
-    	    
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-        
+		structures.valuesInDashboard1 = new HashMap<>();
+
+		String sqlQuery = "SELECT name, ID, byte FROM data WHERE projectName = ?";
+
+		ResultSet rs = db.query(sqlQuery, projectName);
+
+		while (rs.next()) {
+
+			String dataName = rs.getString("name");
+			String id = rs.getString("id");
+			String byteNumber = rs.getString("byte");
+
+			if (!structures.valuesInDashboard1.containsKey(dataName)) {
+				structures.valuesInDashboard1.put(dataName, new ArrayList<>());
+			}
+
+			structures.valuesInDashboard1.get(dataName).add(id);
+			structures.valuesInDashboard1.get(dataName).add(byteNumber);
+			structures.dataQ1Dash1 = new HashMap<>();
+
+		}
+
+		// Sets new project
+		RootViewController.projectObj = new Project(projectName);
+		RootViewController.projectName.setValue(projectName);
+
+		// Closes window
+		Stage stage = (Stage) btnCancel.getScene().getWindow();
+		stage.close();
+
 	}
-	
-	
+
 	/**
 	 * FXML Function
 	 * 
@@ -176,59 +167,55 @@ public class ProjectListController {
 	 */
 	@FXML
 	private void deleteProject(ActionEvent event) {
-		
 
 		if (listProjects.getSelectionModel().getSelectedItem() == null) {
-			GuiUtils.generateAlert(AlertType.WARNING, "Select a item", "You must select a project to delete");
+			GuiUtils.generateAlert(AlertType.WARNING, "Select a item", "You must select a project to be deleted.");
 		} else {
-			
-			//Ask for confirmation.
+
+			// Ask for confirmation.
 			Alert alert = new Alert(AlertType.CONFIRMATION);
 			alert.setTitle("Delete project");
 			alert.setHeaderText("Are you sure you want to delete the project?");
-	
-			Optional<ButtonType> result = alert.showAndWait();
-			
-			if (result.get() == ButtonType.OK){
-				
-				/**
-				 * Delete project from database.
-				 */
-				String itemToDelete = listProjects.getSelectionModel().getSelectedItem().toString();
-	
-				String sql = "DELETE FROM projects WHERE name = ?";
-				String sql2 = "DELETE FROM data WHERE projectName = ?";
-				 try (Connection conn = DbUtils.connect();
-			            PreparedStatement pstmt = conn.prepareStatement(sql)) {
-					 
-			            pstmt.setString(1, itemToDelete);
-			            
-			            pstmt.executeUpdate();
-			 
-			        } catch (SQLException e) {
-			        	GuiUtils.generateAlert(AlertType.ERROR, "Internal error", "Error deleting project.");
-			        }
-				 
-				 try (Connection conn = DbUtils.connect();
-				            PreparedStatement pstmt = conn.prepareStatement(sql2)) {
 
-				            pstmt.setString(1, itemToDelete);
-				            
-				            pstmt.executeUpdate();
-				 
-				        } catch (SQLException e) {
-				            GuiUtils.generateAlert(AlertType.ERROR, "Internal error", "Error deleting project.");
-				        }
-				
-				//Delete item from list
-				listProjects.getItems().remove(listProjects.getSelectionModel().getSelectedItem());
-				
-				GuiUtils.generateAlert(AlertType.INFORMATION, "Success", "Project deleted");
+			Optional<ButtonType> result = alert.showAndWait();
+
+			if (result.isPresent()) {
+
+				if (result.get() == ButtonType.OK) {
+
+					/**
+					 * Delete project from database.
+					 */
+					String itemToDelete = listProjects.getSelectionModel().getSelectedItem().toString();
+
+					String sql = "DELETE FROM projects WHERE name = ?";
+					String sql2 = "DELETE FROM data WHERE projectName = ?";
+
+					try {
+						if (!db.delete(sql, itemToDelete)) {
+							GuiUtils.generateAlert(AlertType.ERROR, "INTERNAL ERROR", "Error deleting project.");
+						}
+					} catch (SQLException e1) {
+						GuiUtils.generateAlert(AlertType.ERROR, "INTERNAL ERROR", "Error executing query");
+					}
+
+					try {
+						if (!db.delete(sql2, itemToDelete)) {
+							GuiUtils.generateAlert(AlertType.ERROR, "INTERNAL ERROR", "Error deleting project.");
+						}
+					} catch (SQLException e) {
+						GuiUtils.generateAlert(AlertType.ERROR, "INTERNAL ERROR", "Error executing query");
+					}
+
+					// Delete item from list
+					listProjects.getItems().remove(listProjects.getSelectionModel().getSelectedItem());
+
+					GuiUtils.generateAlert(AlertType.INFORMATION, "Success", "Project deleted");
 				}
+			}
 		}
 	}
-	
-	
+
 	/**
 	 * FXML Function.
 	 * 
@@ -239,22 +226,21 @@ public class ProjectListController {
 	 */
 	@FXML
 	private void importProject(ActionEvent event) throws IOException {
-		//Load FXML file.
+		// Load FXML file.
 		FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/com/cret/gui/ImportView.fxml"));
 		Parent importDialog = (Parent) fxmlLoader.load();
 		Stage stage = new Stage();
-		//Set icon.
+		// Set icon.
 		stage.getIcons().add(new Image("file:resources/icon.png"));
-		//Set title.
+		// Set title.
 		stage.setTitle("CRET - Import project");
 		stage.setScene(new Scene(importDialog));
 		stage.setResizable(false);
-		//Show window.
+		// Show window.
 		stage.show();
 
-		
 	}
-	
+
 	/**
 	 * FXML funciton.
 	 * 
@@ -265,22 +251,22 @@ public class ProjectListController {
 	 */
 	@FXML
 	private void exportProject(ActionEvent event) throws IOException {
-		
-		//Load FXML file.
+
+		// Load FXML file.
 		FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/com/cret/gui/ExportView.fxml"));
 		Parent exportDialog = (Parent) fxmlLoader.load();
 		Stage stage = new Stage();
-		//Set icon.
+		// Set icon.
 		stage.getIcons().add(new Image("file:resources/icon.png"));
-		//Set title.
+		// Set title.
 		stage.setTitle("CRET - Export project");
 		stage.setScene(new Scene(exportDialog));
 		stage.setResizable(false);
-		//Show window.
+		// Show window.
 		stage.show();
-		
+
 	}
-	
+
 	/**
 	 * FXML function.
 	 * 
@@ -288,20 +274,20 @@ public class ProjectListController {
 	 */
 	@FXML
 	private void initialize() {
-		
+
 		db = new Database("cret.db");
-		
+
 		String query = "SELECT name FROM projects";
-		
+
 		ResultSet rs = db.query(query);
-		
+
 		try {
-			while(rs.next()) {
+			while (rs.next()) {
 				listProjects.getItems().add(rs.getString("name"));
 			}
 		} catch (SQLException e) {
 
-			GuiUtils.generateAlert(AlertType.ERROR, "Error getting data", "Error retrieveing data from database");
+			GuiUtils.generateAlert(AlertType.ERROR, "INTERNAL ERROR", "Error retrieveing data from database");
 		}
 
 	}
